Fixed license header, reduced duplicate code, reduced flaky tests
